# Fix h264_v4l2m2m acceleration in Raspberry Pi 4