@vacafc 2022-02-12
are you using a 64 bit OS on your Raspberry Pi 4? after following all the steps and attempting to run the container, the output is this:
```
WARNING: The requested image's platform (linux/arm64/v8) does not match the detected host platform (linux/arm/v7) and no specific platform was requested
standard_init_linux.go:228: exec user process caused: exec format error```