@aolszowka 2022-02-12
@vacafc Yes I am

```
ubuntu@rbpi4b:~$ lsb_release -a && uname -m
No LSB modules are available.
Distributor ID: Ubuntu
Description:    Ubuntu 21.10
Release:        21.10
Codename:       impish
aarch64
```