@abdulmueid 2022-02-13
Regarding performance, decoding performance could be the botttleneck (depending on the video resolution) rather than encoding.

Try adding `dtoverlay=rpivid-v4l2` to /boot/config.txt and passing /dev/video19 to the container. That should enable hardware accelerated decoding.

Read More:
https://forums.raspberrypi.com/viewtopic.php?t=283301&start=50
https://forum.kodi.tv/showthread.php?tid=355903&page=2