@dfunkt 2022-02-21
> > The performance is still pretty bad, making me wonder if this is even being hardware accelerated at all?
> 
> Unlike HWA on intel, amd and nvidia gpus, the [decoder and filters in V4L2](https://github.com/jc-kynesim/rpi-ffmpeg) are not finished yet. Applying CPU based deinterlacing and scaling filters will slow down the transcoding.

Well, if I understood corectly, the deinterlacing filter seems to be implemented now: https://github.com/jc-kynesim/rpi-ffmpeg/commit/d1f4c1227b5520e4b7b21e3f57a187ddfdc05575