@nyanmisaka 2022-02-22
> Well, if I understood corectly, the deinterlacing filter seems to be implemented now: [jc-kynesim/rpi-ffmpeg@d1f4c12](https://github.com/jc-kynesim/rpi-ffmpeg/commit/d1f4c1227b5520e4b7b21e3f57a187ddfdc05575)

I saw it last year but the lacking of `scale_v4l2m2m` filter will also kill the speed.